Everything associated with the game from Eclipse.
I have created a .jar and .exe associated with the game as well as a
low-grade logo and animation and will upload upon request. Take a look
at the code and let me know what parts are difficult to understand. This
is preliminary as I plan to go back and comment as much as necessary,
but try to get some level of association with the code and my coding
style and maybe catch some redundancies. I hope to salvage as much of
the existing code as possible but it may be difficult considering that I
wrote it all over 6 months ago. Let me know what you think. Cheers!

- Cross